split symbol index into dedicated microcrate
