+++ Cargo.toml
@@ -51,6 +51,7 @@ members = [
     "crates/perl-position-tracking",
     "crates/perl-symbol-types",
     "crates/perl-symbol-cursor",
+    "crates/perl-symbol-index",
     "crates/perl-symbol-table",
     "crates/perl-module-boundary",
     "crates/perl-module-token-core",
@@ -127,6 +128,7 @@ allow = [
     "perl-incremental-parsing",
     "perl-symbol-types",
 "perl-symbol-cursor",
+"perl-symbol-index",
 "perl-symbol-table",
     "perl-uri",
     "perl-workspace-index-slo",
@@ -233,6 +235,7 @@ perl-tokenizer = { path = "crates/perl-tokenizer", version = "0.10.0" }
 perl-position-tracking = { path = "crates/perl-position-tracking", version = "0.10.0" }
 perl-symbol-types = { path = "crates/perl-symbol-types", version = "0.10.0" }
 perl-symbol-cursor = { path = "crates/perl-symbol-cursor", version = "0.10.0" }
+perl-symbol-index = { path = "crates/perl-symbol-index", version = "0.10.0" }
 perl-symbol-table = { path = "crates/perl-symbol-table", version = "0.10.0" }
 perl-module-boundary = { path = "crates/perl-module-boundary", version = "0.10.0" }
 perl-module-token-core = { path = "crates/perl-module-token-core", version = "0.10.0" }
