+++ crates/perl-symbol-index/README.md
@@ -1,0 +1,3 @@
+# perl-symbol-index
+
+Single-responsibility symbol indexing microcrate for prefix and fuzzy symbol matching.
