+++ crates/perl-symbol-index/src/lib.rs
@@ -1,0 +1,168 @@
+//! Symbol indexing helpers for prefix and fuzzy matching.
+
+#![deny(unsafe_code)]
+#![warn(rust_2018_idioms)]
+#![warn(missing_docs)]
+#![warn(clippy::all)]
+
+use std::collections::HashMap;
+
+/// Symbol index for fast lookups.
+///
+/// Supports both prefix and fuzzy matching using a trie and inverted index.
+pub struct SymbolIndex {
+    /// Trie structure for prefix matching
+    trie: SymbolTrie,
+    /// Inverted index for fuzzy matching
+    inverted_index: HashMap<String, Vec<String>>,
+}
+
+/// Trie data structure for efficient prefix matching
+struct SymbolTrie {
+    /// Child nodes indexed by character
+    children: HashMap<char, Box<SymbolTrie>>,
+    /// Symbols stored at this node
+    symbols: Vec<String>,
+}
+
+impl Default for SymbolIndex {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl SymbolIndex {
+    /// Create a new empty symbol index
+    pub fn new() -> Self {
+        Self { trie: SymbolTrie::new(), inverted_index: HashMap::new() }
+    }
+
+    /// Add a symbol to the index.
+    ///
+    /// Indexes the symbol for both prefix and fuzzy matching.
+    pub fn add_symbol(&mut self, symbol: String) {
+        self.trie.insert(&symbol);
+
+        let tokens = Self::tokenize(&symbol);
+        for token in tokens {
+            self.inverted_index.entry(token).or_default().push(symbol.clone());
+        }
+    }
+
+    /// Search symbols with prefix.
+    ///
+    /// Returns all symbols starting with the given prefix.
+    pub fn search_prefix(&self, prefix: &str) -> Vec<String> {
+        self.trie.search_prefix(prefix)
+    }
+
+    /// Fuzzy search symbols.
+    ///
+    /// Returns symbols matching any of the tokenized query words, sorted by relevance.
+    pub fn search_fuzzy(&self, query: &str) -> Vec<String> {
+        let tokens = Self::tokenize(query);
+        let mut results = HashMap::new();
+
+        for token in tokens {
+            if let Some(symbols) = self.inverted_index.get(&token) {
+                for symbol in symbols {
+                    *results.entry(symbol.clone()).or_insert(0) += 1;
+                }
+            }
+        }
+
+        let mut sorted: Vec<_> = results.into_iter().collect();
+        sorted.sort_by(|(_, a), (_, b)| b.cmp(a));
+
+        sorted.into_iter().map(|(symbol, _)| symbol).collect()
+    }
+
+    fn tokenize(s: &str) -> Vec<String> {
+        let mut tokens = Vec::new();
+        let mut current = String::new();
+        let mut prev_upper = false;
+
+        for ch in s.chars() {
+            if ch.is_uppercase() && !prev_upper && !current.is_empty() {
+                tokens.push(current.to_lowercase());
+                current = String::new();
+            }
+
+            if ch.is_alphanumeric() {
+                current.push(ch);
+                prev_upper = ch.is_uppercase();
+            } else if !current.is_empty() {
+                tokens.push(current.to_lowercase());
+                current = String::new();
+                prev_upper = false;
+            }
+        }
+
+        if !current.is_empty() {
+            tokens.push(current.to_lowercase());
+        }
+
+        tokens
+    }
+}
+
+impl SymbolTrie {
+    fn new() -> Self {
+        Self { children: HashMap::new(), symbols: Vec::new() }
+    }
+
+    fn insert(&mut self, symbol: &str) {
+        let mut node = self;
+
+        for ch in symbol.chars() {
+            node = node.children.entry(ch).or_insert_with(|| Box::new(SymbolTrie::new()));
+        }
+
+        node.symbols.push(symbol.to_string());
+    }
+
+    fn search_prefix(&self, prefix: &str) -> Vec<String> {
+        let mut node = self;
+
+        for ch in prefix.chars() {
+            match node.children.get(&ch) {
+                Some(child) => node = child,
+                None => return Vec::new(),
+            }
+        }
+
+        let mut results = Vec::new();
+        Self::collect_all(node, &mut results);
+        results
+    }
+
+    fn collect_all(node: &SymbolTrie, results: &mut Vec<String>) {
+        results.extend(node.symbols.clone());
+
+        for child in node.children.values() {
+            Self::collect_all(child, results);
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_symbol_index() {
+        let mut index = SymbolIndex::new();
+
+        index.add_symbol("calculate_total".to_string());
+        index.add_symbol("calculate_average".to_string());
+        index.add_symbol("get_user_name".to_string());
+
+        let results = index.search_prefix("calc");
+        assert_eq!(results.len(), 2);
+        assert!(results.contains(&"calculate_total".to_string()));
+        assert!(results.contains(&"calculate_average".to_string()));
+
+        let results = index.search_fuzzy("user name");
+        assert!(results.contains(&"get_user_name".to_string()));
+    }
+}
